requirements.txt
pandas
python-dotenv
requests
azure-storage-blob
